added images to the dashboards, and created home page for the dataviewers where they are displayed
Former-commit-id: 31b19805067c18962c86969bf1c5a92ee97ded32

+++ modules/control-panel/src/main/java/com/indracompany/sofia2/controlpanel/controller/dashboard/DashboardController.java
@@ -13,11 +13,14 @@
  */
 package com.indracompany.sofia2.controlpanel.controller.dashboard;
 
+import java.io.IOException;
+import java.io.OutputStream;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -34,6 +37,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.multipart.MultipartHttpServletRequest;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -133,7 +137,7 @@ public class DashboardController {
 
 	@PostMapping(value = { "/create" })
 	public String createDashboard(Model model, @Valid DashboardCreateDTO dashboard, BindingResult bindingResult,
-			RedirectAttributes redirect) {
+			MultipartHttpServletRequest request, RedirectAttributes redirect) {
 		if (bindingResult.hasErrors()) {
 			utils.addRedirectMessage("dashboard.validation.error", redirect);
 			return "redirect:/dashboards/create";
@@ -150,8 +154,8 @@ public class DashboardController {
 		}
 	}
 
-	@PutMapping(value = { "/dashboardconf/{id}" })
-	public String saveUpdateDashboard(Model model, @Valid DashboardCreateDTO dashboard, @PathVariable("id") String id,
+	@PostMapping(value = { "/dashboardconf/{id}" })
+	public String saveUpdateDashboard(@PathVariable("id") String id, DashboardCreateDTO dashboard,
 			BindingResult bindingResult, RedirectAttributes redirect) {
 		if (bindingResult.hasErrors()) {
 			utils.addRedirectMessage("dashboard.validation.error", redirect);
@@ -186,6 +190,12 @@ public class DashboardController {
 			dashBDTO.setId(id);
 			dashBDTO.setIdentification(dashboard.getIdentification());
 			dashBDTO.setDescription(dashboard.getDescription());
+			if (null != dashboard.getImage()) {
+				dashBDTO.setHasImage(Boolean.TRUE);
+			} else {
+				dashBDTO.setHasImage(Boolean.FALSE);
+			}
+			// dashBDTO.setImage(dashboard.getImage().getBytes());
 			dashBDTO.setPublicAccess(dashboard.isPublic());
 			List<DashboardUserAccess> userAccess = dashboardService.getDashboardUserAccesses(id);
 			if (userAccess != null && userAccess.size() > 0) {
@@ -270,8 +280,30 @@ public class DashboardController {
 
 	@DeleteMapping("/{id}")
 	public String delete(Model model, @PathVariable("id") String id) {
-		this.dashboardService.deleteDashboard(id, utils.getUserId());
+		if (dashboardService.hasUserEditPermission(id, utils.getUserId())) {
+			this.dashboardService.deleteDashboardAccess(id, utils.getUserId());
+			this.dashboardService.deleteDashboard(id, utils.getUserId());
+		}
 		return "redirect:/dashboards/list/";
+	}
+
+	@RequestMapping(value = "/{id}/getImage")
+	public void showImg(@PathVariable("id") String id, HttpServletResponse response) {
+		byte[] buffer = this.dashboardService.getImgBytes(id);
+		if (buffer.length > 0) {
+			OutputStream output = null;
+			try {
+				output = response.getOutputStream();
+				response.setContentLength(buffer.length);
+				output.write(buffer);
+			} catch (Exception e) {
+			} finally {
+				try {
+					output.close();
+				} catch (IOException e) {
+				}
+			}
+		}
 	}
 
 }
